third_party/blink/web_tests/external/wpt/service-workers/service-worker/registration-updateviacache.https-expected.txt

This is a testharness.js-based test.
[FAIL] updateViaCache is not updated if register() rejects
  assert_equals: after update attempt expected "imports" but got "none"
Harness: the test ran to completion.
